+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  - add 2d interpolation
  - add biliniar interpolation strategy
  - rename `Strategy` to `Interp1DStrategy` and `StrategyBuilder` to `Interp1DStrategyBuilder`
+ - make extrapolate filed of `Linear` private add `extrapolate(bool)` method instead.
 
 # 0.2.1
  - change interp_array such that it can be called with any